docs: add readme & fix build error
