+++ web/src/app/components/title.tsx
@@ -1,4 +1,5 @@
 "use client";
+import { getSearchUrl } from "@/app/utils/get-search-url";
 import { RefreshCcw } from "lucide-react";
 import { nanoid } from "nanoid";
 import { useRouter } from "next/navigation";
@@ -16,9 +17,7 @@ export const Title = ({ query }: { query: string }) => {
       <div className="flex-none">
         <button
           onClick={() => {
-            router.push(
-              `/search.html?q=${encodeURIComponent(query)}&rid=${nanoid()}`,
-            );
+            router.push(getSearchUrl(encodeURIComponent(query), nanoid()));
           }}
           type="button"
           className="rounded flex gap-2 items-center bg-transparent px-2 py-1 text-xs font-semibold text-blue-500 hover:bg-zinc-100"
