+++ web/src/app/utils/get-search-url.ts
@@ -1,0 +1,5 @@
+export const getSearchUrl = (query: string, search_uuid: string) => {
+  const prefix =
+    process.env.NODE_ENV === "production" ? "/search.html" : "/search";
+  return `${prefix}?q=${encodeURIComponent(query)}&rid=${search_uuid}`;
+};
